Rename parsers to attr_parser
